+++ CHANGELOG.md
@@ -178,6 +178,10 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
   [Kra Larivain](https://github.com/olarivain)
   [#2532](https://github.com/CocoaPods/CocoaPods/pull/2532)
 
+* Proper scoping of header search paths to the target platform.  
+  [Michael Melanson](https://github.com/michaelmelanson)
+  [#1249](https://github.com/CocoaPods/CocoaPods/issues/1249)
+
 
 ## 0.34.1